// src/app/service/model/BlogDTO.ts
/**
 * SpongeblogSP API
 * Spongeblog blogging API
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface BlogDTO {
    links?: Array<models.Link>;

    /**
     * Name of the Blog.
     */
    name?: string;

    /**
     * Settings map of the Blog.
     */
    settings?: { [key: string]: string; };

    /**
     * Status of the Blog.
     */
    status?: BlogDTO.StatusEnum;

}
export namespace BlogDTO {
    export enum StatusEnum {
        ACTIVE = <any> 'ACTIVE',
        DISABLED = <any> 'DISABLED',
        MAINTENANCE = <any> 'MAINTENANCE'
    }
}